Add the insert head function
First, check the input head, malloc a space for the new element,
after that create a space for the content of the element, if one of them
fail to create a space then return false. After successfully created,
add element to the list with the linux kernal api list_add and return
true.