Add the q_delete_dup function
Inital a list head to contain the elements needed to delete from
original list, which is called pending list. Check the content of element in the for loop, if the
first content and second content cannot match, then check whether there
the pointer is point to itself, if so then assign pointer to next
element, if not then split the list out of the original list, and add it
to the pending list. At last, release the pending list.