Add q_merge function and modify q_merge2 function
Use the q_merge2 to implement the q_merge function. At first I find some
problem when I used q_merge2 before modified. When I merge in the q_merge2
the head of the list which merge into the other list will not be reset
and it will cause error when using this head in the future, so I used
list_spice_init instead of list_spice, and it will solve this problem.