Modify the response message to the browser
After openned the web mode in this project, user can use the browser to
control the queue as command line, however when using the browser I
need to response with the favicon.ico, in the previous code didn't
contain it. I modfity the response content according to homework
discription, but it came out wrong output, it will do the command twice.